fix editor comment, uncomment and indent
previously the selection goes awry in the following cases:
* uncommenting without a comment
* change indent when original tabs become spaces

selection ending at the start of a line will now exclude the line from being indented.

also re-indenting after unindenting now will always be identity. previously if the selection started at the beginning of a line and it ended after an indent the final selection ending on the start of the line causes it to not include the line. now, we force at least the first character of the last line unindented to be selected.